# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Educational Content Generator API",
    "description": "AI-powered educational content generation system built with Strands Agents",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "post": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "Add Document",
        "description": "Add a document to the AWS Bedrock Knowledge Base.\n\nThe document is uploaded to S3 and will be ingested into the Knowledge Base\nwhen sync_data_source is called.",
        "operationId": "add_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "List Documents",
        "description": "List documents in the Knowledge Base S3 bucket.",
        "operationId": "list_documents_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/sync/{job_id}": {
      "get": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "Get Sync Status",
        "description": "Get the status of an ingestion job.",
        "operationId": "get_sync_status_api_v1_knowledge_sync__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/retrieve": {
      "post": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "Retrieve Documents",
        "description": "Retrieve relevant documents from the Knowledge Base using semantic search.",
        "operationId": "retrieve_documents_api_v1_knowledge_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/rag": {
      "post": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "Retrieve And Generate",
        "description": "Retrieve relevant documents and generate a response using RAG.\n\nThis combines retrieval from the Knowledge Base with generation\nusing a foundation model for question answering over your documents.",
        "operationId": "retrieve_and_generate_api_v1_knowledge_rag_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveAndGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents/{document_id}": {
      "delete": {
        "tags": [
          "knowledge-base",
          "knowledge"
        ],
        "summary": "Delete Document",
        "description": "Delete a document from the Knowledge Base S3 bucket.\n\nThe document will be removed from the Knowledge Base on the next sync.",
        "operationId": "delete_document_api_v1_knowledge_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "educational_content",
              "title": "Content Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddDocumentRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "default": {

            }
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "educational_content"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "content"
        ],
        "title": "AddDocumentRequest",
        "description": "Request model for adding a document to the Knowledge Base."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetrieveAndGenerateRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "default": "amazon.nova-premier-v1:0"
          },
          "max_results": {
            "type": "integer",
            "title": "Max Results",
            "default": 5
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveAndGenerateRequest",
        "description": "Request model for RAG queries."
      },
      "RetrieveRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "max_results": {
            "type": "integer",
            "title": "Max Results",
            "default": 5
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveRequest",
        "description": "Request model for retrieving documents."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}